requirements.txt
numpy==1.23.1
Pillow==9.2.0
google-cloud-storage==2.5.0
kaggle==1.5.12
pandas==1.4.3
joblib==1.1.0
tqdm==4.64.0

